# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Trento Apartment Service API
  description: |
    API REST per la piattaforma di ricerca e gestione di appartamenti e annunci a Trento.

    Gli endpoint protetti richiedono un token JWT nell'header `Authorization: Bearer <token>`.
    Il token viene rilasciato al login o alla registrazione e ha una durata di 7 giorni (configurabile via `JWT_EXPIRES_IN`).

    **Ruoli utente disponibili:**
    - `utente base` *(default)* — cerca annunci
    - `utente verificato` — cerca annunci (account verificato)
    - `amministratore` — gestisce appartamenti e annunci; può essere privato o agenzia (con partita IVA)
    - `dipendente comune` — operatore del comune; accede alla dashboard statistica TAS Data
    - `inquilino` — inquilino con contratto attivo
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Server di sviluppo locale
security: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT ottenuto al login o alla registrazione. Payload contiene `sub` (ID utente), `email`, `ruolo`.
  schemas:
...
    Bolletta:
      type: object
      properties:
        _id:
          type: string
          example: 665f1a2b3c4d5e6f7a8b9c30
        idAppartamento:
          type: string
          example: 665f1a2b3c4d5e6f7a8b9c0d
        utenza:
          type: string
          enum: [luce, gas, acqua, elettricità]
          example: elettricità
        periodoInizio:
          type: string
          format: date
          example: "2026-01-01"
        periodoFine:
          type: string
          format: date
          example: "2026-03-31"
        importo:
          type: number
          minimum: 0
          example: 85.5
        pagata:
          type: boolean
          default: false
          example: false
        pdfNomeFile:
          type: string
          nullable: true
          description: Nome del PDF allegato, se presente
          example: bolletta-gennaio.pdf
...
    ErroreGenerico:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Errore interno del server
    ErroreAutenticazione:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Token non fornito. Usa Authorization: Bearer <token>"
    ErroreAutorizzazione:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Accesso non autorizzato

# ---------------------------------------------------------------------------
# PATH
# ---------------------------------------------------------------------------
paths:
...
  /api/v1/bollette/{appartamentoId}:
    post:
      tags: [Bollette]
      summary: Carica una nuova bolletta (con PDF opzionale)
      description: Solo l'amministratore proprietario dell'appartamento. Il PDF, se presente, deve essere in formato application/pdf.
      operationId: caricaBolletta
      security:
        - bearerAuth: []
      parameters:
        - name: appartamentoId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [utenza, periodoInizio, periodoFine, importo]
              properties:
                utenza:
                  type: string
                  enum: [luce, gas, acqua, elettricità]
                periodoInizio:
                  type: string
                  format: date
                periodoFine:
                  type: string
                  format: date
                importo:
                  type: number
                pdf:
                  type: string
                  format: binary
                  description: File PDF della bolletta (opzionale, max 10 MB)
      responses:
        "201":
          description: Bolletta creata
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Bolletta'
        "400":
          description: Campi mancanti o file non PDF
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "403":
          description: Non sei l'amministratore proprietario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreAutorizzazione'
  /api/v1/appartamenti/{appId}/bollette:
    get:
      tags: [Bollette]
      summary: Lista bollette di un appartamento
      description: |
        Restituisce le bollette dell'appartamento. Accessibile dall'amministratore proprietario
        e dall'inquilino con contratto attivo o in chiusura; il ruolo e l'autorizzazione sono
        determinati dal token JWT. Include i dati aggregati per i grafici dei consumi.
      operationId: getBolletteAppartamento
      security:
        - bearerAuth: []
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
          description: MongoDB ObjectId dell'appartamento
      responses:
        "200":
          description: Elenco bollette e dati per i grafici
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Bolletta'
                  grafici:
                    type: object
                    description: Dataset pronto per i grafici (labels e datasets per utenza)
        "401":
          description: Token JWT assente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreAutenticazione'
        "403":
          description: Non autorizzato (non proprietario o senza contratto attivo per l'appartamento)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreAutorizzazione'
...
  /api/v1/bollette/{bollettaId}:
    delete:
      tags: [Bollette]
      summary: Elimina una bolletta
      operationId: eliminaBolletta
      security:
        - bearerAuth: []
      parameters:
        - name: bollettaId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bolletta eliminata
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Bolletta eliminata
        "404":
          description: Bolletta non trovata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
...